+++ NEWS.rst
@@ -6,7 +6,7 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Patch: Add support for writing a .nojekyll file in the push step.
 
 1.0.2
 -----